beter init scheme and no dropout
This init scheme initializes the output weights/biases to be way smaller, since the outputs tend to only vary between -0.49 and 0.51 or thereabouts. This prevents the network from having to initially bias itself a great deal (which it does by adding up a ton of output states, I assume, since it doesn't have time to move the bias alone).